+++ docs/conf.py
@@ -19,6 +19,7 @@ import id_translation
 
 type_modules = (
     "id_translation.types",
+    "id_translation.fetching",
     "id_translation.fetching.types",
     # rics
     "rics.mapping.types",
@@ -38,6 +39,9 @@ def callback(_app, _env, node, _contnode):  # noqa
         #   Special hack for factory.py, which is a public module. And for some
         #   reason that breaks. I've no idea why :')
         reftarget = f"id_translation.types.{reftarget}"
+
+    if reftarget == "id_translation.fetching._sql_fetcher.StatementType":
+        reftarget = "id_translation.fetching.SqlFetcher.StatementType"
 
     for m in type_modules:
         if reftarget.startswith(m):